Complete search api endpoint
Adds API endpoint /api/v1/search which will take query parameter
of searchName and run a database query on foods to find matching
foods. Search is case insensitive and looks for occurrences of
searchName parameter throughout the name of a food. Tests are
currently passing for both happy and sad path. In the event of
no search results, an empty array is returned.