Complete delete food from meal endpoint
Modifies diary meals endpoint to no longer return food_id;
this was found to be unnecessary for the purpose of the
end goal. Endpoint added at DELETE /api/v1/meals/:id.
Deletes selected meal item; this results in the food
being preserved in the foods table; however, the
association between the food and the diary is destroyed.
Tests are all currently passing.